+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfbcvk2z79s7bnphveex5em6"
+        "conv_01kfbeeg115zpsbsdrbaebs4zz"
     )
     r"""Unique conversation identifier with `conv_` prefix."""