+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFBCVKKB04MW4M0ADD7Q07S7"
+        "01KFBEEGHDBSZHVMPDW803T18D"
     )
     r"""The unique identifier of the data source"""