+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFBCVKK8XRXWD83770YJ8FQW"
+        "01KFBEEGH0QBJK93GAGZPEJ2XY"
     )
     r"""The unique identifier of the data source"""